[minor] Move StreamRecordMatchers to util package
